backend/requirements.txt
fastapi
uvicorn
supabase
google-genai
python-dotenv
pydantic
easyocr
PyJWT
pytest
httpx
